will not overwrite existing model files

+++ lib/modeller/printer.rb
@@ -3,8 +3,12 @@ module Modeller
     @@path = "#{Rails.root || "."}/app/models/" # outside the Rails env?
     
     def self.print_model(model)
-      File.open("#{@@path}#{model.filename}", "w") do |f|
-        f.write(model.to_s)
+      if File.exists?("#{@@path}#{model.filename}") 
+        puts "Warning: #{@@path}#{model.filename} exists. Not overwriting."
+      else
+        File.open("#{@@path}#{model.filename}", "w") do |f|
+          f.write(model.to_s)
+        end
       end
     end
   end
